update the userRessource
change userId to Long and add proper swagger description

+++ src/main/java/org/obs/app/repository/UserRepository.java
@@ -14,7 +14,7 @@ public class UserRepository {
     @Inject
     JPAStreamer jpaStreamer;
 
-    public Optional<User> getUser(Integer userId){
+    public Optional<User> getUser(Long userId){
         return jpaStreamer.stream(User.class)
                 .filter(User$.id.equal(userId))
                 .findFirst();
